ting/happyface/requirements.txt
pillow
requests
transformers
torch
numpy
opencv-python
